Declare generic EntityStore as parameter type
Store::getEntityStore() (which is where
ImportEntities::newEntityImporter() gets the entity store from) has
never promised to return more than a generic EntityStore, and with
Wikibase commit 91a07a3901 (change I84661212a6), it has started to
return a TypeDispatchingEntityStore instead of a WikiPageEntityStore. As
far as I can tell, we don’t need anything specific to
WikiPageEntityStore, so adjusting the parameter type should be enough to
fix this problem.